Store values in the float_column as the double type

+++ Catalog/VOTableCarrier.cc
@@ -96,9 +96,10 @@ void VOTableCarrier::FillTdValues(int column_index, std::string value) {
         }
     } else if (_fields[column_index].datatype == "float") {
         try {
-            _float_vectors[column_index].push_back(std::stof(value));
+            // TODO: Due to the precision problem for the float type, we convert it as a double type instead.
+            _float_vectors[column_index].push_back(std::stod(value));
         } catch (...) {
-            _float_vectors[column_index].push_back(std::numeric_limits<float>::quiet_NaN());
+            _float_vectors[column_index].push_back(std::numeric_limits<double>::quiet_NaN());
         }
     } else if (_fields[column_index].datatype == "double") {
         try {
@@ -231,7 +232,7 @@ void VOTableCarrier::GetHeadersAndData(CARTA::OpenCatalogFileAck& open_file_resp
                 // Fill the long long column index
                 header->set_data_type_index(columns_data->ll_column_size() - 1);
             } else if (_float_vectors.count(column_index)) {
-                std::vector<float>& ref_column_data = _float_vectors[column_index];
+                std::vector<double>& ref_column_data = _float_vectors[column_index];
                 // Add a float column
                 auto float_columns_data = columns_data->add_float_column();
                 // Fill float column elements
@@ -477,7 +478,7 @@ void VOTableCarrier::GetFilteredData(
                 ll_column->add_ll_column(ll_vector.second[row]);
             }
         }
-        for (std::pair<int, std::vector<float>> float_vector : _float_vectors) {
+        for (std::pair<int, std::vector<double>> float_vector : _float_vectors) {
             int column_index = float_vector.first;
             if (hided_column_indices.find(column_index) == hided_column_indices.end()) {
                 int data_type_index = column_to_data_type_index[column_index];
